Refactor Parameters sub-package to reduce code duplication.
