+++ test/Parameters/test_parameters.py
@@ -14,19 +14,22 @@ else:
     has_tleap = False
 
 @pytest.mark.skipif(has_tleap is False, reason="Requires tLEaP to be installed.")
-def test_disulphide():
+@pytest.mark.parametrize(
+        "ff", BSS.Parameters.amberProteinForceFields()
+)
+def test_disulphide(ff):
     """Test parameterisation in the presence of disulphide bridges."""
 
     # Load the example molecule.
     molecule = BSS.IO.readMolecules("test/input/molecules/4LYT_Fixed.pdb")[0]
 
-    # Try to parameterise with ff14SB. If working, this should auto-detect
-    # disulphide bonds and add the appropriate bond records to the tLEaP input
-    # script.
-    molecule = BSS.Parameters.ff14SB(molecule).getMolecule()
+    # Try to parameterise with the named force field. If working, this should
+    # auto-detect disulphide bonds and add the appropriate bond records to the
+    # tLEaP input script.
+    molecule = getattr(BSS.Parameters, ff)(molecule).getMolecule()
 
     # Check that we actually generate records for four disulphide bonds.
-    bonds = BSS.Parameters.Protocol._protocol._get_disulphide_bonds(molecule._sire_object)
+    bonds = BSS.Parameters._Protocol.AmberProtein._get_disulphide_bonds(molecule._sire_object)
     assert len(bonds) == 4
 
     # Check that the bond parameters are present in the molecule.
